+++ src/components/LocationAutocomplete/LocationAutocomplete.scss
@@ -1,4 +1,4 @@
-.LocationAutocompleteRenderer {
+.LocationAutocomplete {
     &__poweredByGoogleImage {
         position: absolute;
         right: var(--spacing-4x);
