--- src/components/LocationAutocomplete/LocationAutocompleteRenderer/index.js
@@ -1,1 +1,0 @@
-export { default } from './LocationAutocompleteRenderer';
