+++ src/components/LocationAutocomplete/index.js
@@ -1,2 +1,4 @@
+export {
+    default as LocationAutocompleteWithGoogleLoader,
+} from './LocationAutocompleteWithGoogleLoader';
 export { default as LocationAutocomplete } from './LocationAutocomplete';
-export { default as LocationAutocompleteRenderer } from './LocationAutocompleteRenderer';
